Add input feature and adjust pie chart UI

+++ app/src/main/java/com/example/piechart/ChartEntry.kt
@@ -3,5 +3,5 @@ package com.example.piechart
 // Represents a single category and its associated value in the chart
 data class ChartEntry(
     val category: String,  // The name of the category
-    val value: Float       // The value (amount) for the category
+    val value: Double       // The value (amount) for the category
 )
